changed short to int in Function.evaluate() method signature
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@881704 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/record/formula/eval/ConcatEval.java
@@ -30,7 +30,7 @@ public final class ConcatEval implements Function {
 		// enforce singleton
 	}
 
-	public ValueEval evaluate(ValueEval[] args, int srcRow, short srcCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcRow, int srcCol) {
 		if(args.length != 2) {
 			return ErrorEval.VALUE_INVALID;
 		}
