+++ src/java/org/apache/poi/hssf/record/formula/eval/RangeEval.java
@@ -32,7 +32,7 @@ public final class RangeEval implements Function {
 		// enforces singleton
 	}
 
-	public ValueEval evaluate(ValueEval[] args, int srcRow, short srcCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcRow, int srcCol) {
 		if(args.length != 2) {
 			return ErrorEval.VALUE_INVALID;
 		}
