+++ src/java/org/apache/poi/hssf/record/formula/functions/Find.java
@@ -36,7 +36,7 @@ import org.apache.poi.hssf.record.formula.eval.ValueEval;
  */
 public final class Find extends TextFunction {
 
-	protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, short srcCellCol)
+	protected ValueEval evaluateFunc(ValueEval[] args, int srcCellRow, int srcCellCol)
 			throws EvaluationException {
 
 		int nArgs = args.length;
