+++ src/java/org/apache/poi/hssf/record/formula/functions/Match.java
@@ -66,7 +66,7 @@ import org.apache.poi.hssf.record.formula.functions.LookupUtils.ValueVector;
 public final class Match implements Function {
 
 
-	public ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 
 		double match_type = 1; // default
 
@@ -154,7 +154,7 @@ public final class Match implements Function {
 
 
 
-	private static double evaluateMatchTypeArg(ValueEval arg, int srcCellRow, short srcCellCol)
+	private static double evaluateMatchTypeArg(ValueEval arg, int srcCellRow, int srcCellCol)
 			throws EvaluationException {
 		ValueEval match_type = OperandResolver.getSingleValue(arg, srcCellRow, srcCellCol);
