+++ src/java/org/apache/poi/hssf/record/formula/functions/Not.java
@@ -32,7 +32,7 @@ import org.apache.poi.hssf.record.formula.eval.ValueEval;
  */
 public final class Not implements Function {
 
-	public ValueEval evaluate(ValueEval[] args, int srcCellRow, short srcCellCol) {
+	public ValueEval evaluate(ValueEval[] args, int srcCellRow, int srcCellCol) {
 		if (args.length != 1) {
 			return ErrorEval.VALUE_INVALID;
 		}
